backend/requirements.txt
# FastAPI & ASGI
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# Database
motor>=3.4.0
beanie>=1.26.0


# Caching & Task Queue
redis==5.0.4
celery==5.4.0
flower==2.0.1

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
email-validator==2.1.1
slowapi==0.1.9

# Object Storage
minio==7.2.7
firebase-admin==6.5.0

# ML / Data Science
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
shap==0.45.0
imbalanced-learn==0.12.2

# Deep Learning
torch==2.3.0
torchvision==0.18.0
transformers<4.41.0
tokenizers==0.19.1
captum==0.7.0
onnx==1.16.0
onnxruntime==1.18.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
albumentations==1.4.7

# Data Profiling
# ydata-profiling (incompatible with Python 3.14)
ydata-profiling==4.8.3

# FLAML AutoML (lightweight fallback)
flaml==2.2.0

# Utilities
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
boto3==1.34.100
aiofiles==23.2.1
python-slugify==8.0.4
structlog==24.2.0
prometheus-fastapi-instrumentator==6.1.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.4

# ONNX export helpers
optimum==1.19.2
